+++ packages/components/src/Overlay/Popover.test.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable testing-library/no-node-access */
 import { cleanup, screen } from '@testing-library/react';
 import React from 'react';
+import { vi } from 'vitest';
 import { Theme, cva } from '@marigold/system';
 import { Button } from '../Button';
 import { Text } from '../Text';
@@ -21,6 +22,7 @@ const theme: Theme = {
     Dialog: {
       closeButton: cva('p-3'),
       container: cva('p-2'),
+      title: cva(),
       header: cva(),
       content: cva(),
       actions: cva(),
@@ -38,7 +40,7 @@ const { render } = setup({ theme });
  */
 
 const mockMatchMedia = (matches: string[]) =>
-  jest.fn().mockImplementation(query => ({
+  vi.fn().mockImplementation(query => ({
     matches: matches.includes(query),
   }));