+++ packages/components/src/Pagination/NavigationButton.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import { KeyboardEvent, ReactNode, useRef } from 'react';
 import { useButton } from '@react-aria/button';
 import { cn, useClassNames } from '@marigold/system';
 
@@ -7,14 +7,15 @@ interface PaginationButtonProps {
   'aria-label': string;
   isDisabled?: boolean;
   isSelected?: boolean;
-  children: React.ReactNode;
+  children: ReactNode;
   registerRef?: (ref: HTMLButtonElement | null) => void;
   controlLabel?: string;
   position?: 'left' | 'right';
+  onKeyDown?: (event: KeyboardEvent) => void;
 }
 
 export const NavigationButton = (props: PaginationButtonProps) => {
-  const ref = React.useRef<HTMLButtonElement>(null);
+  const ref = useRef<HTMLButtonElement>(null);
   const classNames = useClassNames({
     component: 'Pagination',
   });
@@ -28,13 +29,6 @@ export const NavigationButton = (props: PaginationButtonProps) => {
     position,
     ...rest
   } = props;
-
-  useEffect(() => {
-    if (registerRef) {
-      registerRef(ref.current);
-      return () => registerRef(null);
-    }
-  }, [registerRef]);
 
   return (
     <button