+++ packages/components/src/Pagination/Pagination.stories.tsx
@@ -1,5 +1,6 @@
 import { useState } from '@storybook/preview-api';
 import type { Meta, StoryObj } from '@storybook/react';
+import { expect, fn, userEvent, within } from '@storybook/test';
 import React from 'react';
 import {
   FieldGroup,
@@ -14,6 +15,7 @@ import { Pagination, PaginationProps } from './Pagination';
 
 const meta = {
   title: 'Components/Pagination',
+  component: Pagination,
   argTypes: {
     totalItems: {
       control: {
@@ -35,10 +37,17 @@ const meta = {
         defaultValue: { summary: 'e.g. ["Previous", "Next"]' },
       },
     },
+    defaultPage: {
+      control: {
+        type: 'number',
+      },
+      description: 'The initial page. (uncontrolled)',
+    },
   },
   args: {
     totalItems: 100,
     pageSize: 10,
+    defaultPage: undefined,
   },
 } satisfies Meta<typeof Pagination>;
 
@@ -46,14 +55,70 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Basic: Story = {
-  render: ({ totalItems, pageSize, ...rest }: Partial<PaginationProps>) => (
-    <Pagination
-      {...rest}
-      totalItems={totalItems!}
-      pageSize={pageSize!}
-      defaultPage={5}
-    />
-  ),
+  tags: ['component-test'],
+  play: async ({ canvasElement, step }) => {
+    const canvas = within(canvasElement);
+
+    await step('Select an item from pagination', async () => {
+      const pageButton = canvas.getByLabelText('Page 2');
+
+      await userEvent.click(pageButton);
+
+      await expect(pageButton).toHaveAttribute('data-selected', 'true');
+    });
+
+    await step('Click on the next button', async () => {
+      const nextButton = canvas.getByLabelText('Page next');
+
+      await userEvent.click(nextButton);
+
+      await expect(canvas.getByLabelText('Page 3')).toHaveAttribute(
+        'data-selected',
+        'true'
+      );
+    });
+
+    await step('Click on the previous button', async () => {
+      const prevButton = canvas.getByLabelText('Page previous');
+
+      await userEvent.click(prevButton);
+
+      await expect(canvas.getByLabelText('Page 2')).toHaveAttribute(
+        'data-selected',
+        'true'
+      );
+    });
+
+    await step('use arrow right navigation', async () => {
+      const pageButton = canvas.getByLabelText('Page 2');
+      const nextPageButton = pageButton.nextElementSibling as HTMLElement;
+
+      await userEvent.tab();
+      await userEvent.keyboard('{ArrowRight}');
+
+      await expect(nextPageButton).toHaveFocus();
+      await expect(nextPageButton).toHaveTextContent('3');
+    });
+
+    await step('use arrow left navigation', async () => {
+      const pageButton = canvas.getByLabelText('Page 3');
+      const nextPageButton = pageButton.previousElementSibling as HTMLElement;
+
+      await userEvent.keyboard('{ArrowLeft}');
+
+      await expect(nextPageButton).toHaveFocus();
+      await expect(nextPageButton).toHaveTextContent('2');
+    });
+
+    await step('blur to unfocus pagination', async () => {
+      const pageButton = canvas.getByLabelText('Page 2');
+
+      pageButton.blur();
+
+      await expect(pageButton).not.toHaveFocus();
+      await expect(pageButton).toHaveAttribute('data-selected', 'true');
+    });
+  },
 };
 
 export const Controlled: Story = {
@@ -80,25 +145,27 @@ export const OnePage: Story = {
   parameters: {
     controls: { exclude: ['totalItems', 'pageSize'] },
   },
-  render: args => <Pagination {...args} totalItems={10} pageSize={10} />,
+  args: {
+    totalItems: 10,
+    pageSize: 10,
+  },
 };
 
 export const OneHundredPages: Story = {
-  render: ({ pageSize, ...rest }: Partial<PaginationProps>) => (
-    <Pagination
-      {...rest}
-      totalItems={1000}
-      pageSize={pageSize!}
-      defaultPage={93}
-    />
-  ),
+  args: {
+    totalItems: 1000,
+    defaultPage: 93,
+  },
 };
 
 export const NoData: Story = {
   parameters: {
     controls: { exclude: ['totalItems', 'pageSize'] },
   },
-  render: args => <Pagination {...args} totalItems={0} pageSize={10} />,
+  args: {
+    totalItems: 0,
+    pageSize: 10,
+  },
 };
 
 export const FullScreenSize: Story = {
@@ -218,13 +285,53 @@ export const WithButtonLabels: Story = {
   parameters: {
     controls: { exclude: ['totalItems', 'pageSize'] },
   },
-  render: ({ pageSize, ...rest }: Partial<PaginationProps>) => (
-    <Pagination
-      {...rest}
-      totalItems={100}
-      pageSize={pageSize!}
-      page={5}
-      controlLabels={['Previous', 'Next']}
-    />
-  ),
+  args: {
+    totalItems: 100,
+    defaultPage: 5,
+    controlLabels: ['Previous', 'Next'],
+  },
+};
+
+export const DisabledPreviousButton: Story = {
+  tags: ['component-test'],
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    const previousButton = canvas.getByLabelText('Page previous');
+
+    await userEvent.click(previousButton);
+
+    await expect(previousButton).not.toHaveAttribute('data-selected', 'true');
+    await expect(previousButton).toHaveAttribute('disabled');
+  },
+};
+
+export const DisabledNextButton: Story = {
+  tags: ['component-test'],
+  args: {
+    defaultPage: 9,
+  },
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    const nextButton = canvas.getByLabelText('Page next');
+
+    await userEvent.click(nextButton);
+
+    await expect(nextButton).not.toHaveAttribute('data-selected', 'true');
+    await expect(nextButton).toHaveAttribute('disabled');
+  },
+};
+
+export const UseOnChange: Story = {
+  tags: ['component-test'],
+  args: {
+    onChange: fn(),
+  },
+  play: async ({ canvasElement, args }) => {
+    const canvas = within(canvasElement);
+    const page2Button = canvas.getByLabelText('Page 2');
+
+    await userEvent.click(page2Button);
+
+    await expect(args.onChange).toHaveBeenCalledWith(2);
+  },
 };