+++ packages/components/src/SectionMessage/SectionMessage.tsx
@@ -1,6 +1,7 @@
 import { type ReactNode, useRef, useState } from 'react';
 import { useButton } from '@react-aria/button';
 import { cn, useClassNames } from '@marigold/system';
+import { CloseButton } from '../CloseButton';
 import { SectionMessageContext } from './Context';
 import { SectionMessageContent } from './SectionMessageContent';
 import { SectionMessageTitle } from './SectionMessageTitle';
@@ -129,21 +130,13 @@ export const SectionMessage = ({
           {Icon && <Icon />}
         </div>
         {closeButton && (
-          <button
+          <CloseButton
             {...buttonProps}
             ref={buttonRef}
             aria-label="close"
             className={cn('[grid-area:close]', classNames.close)}
-            onClick={handleClose}
-          >
-            <svg viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
-              <path
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
-              />
-            </svg>
-          </button>
+            onPress={handleClose}
+          />
         )}
         {children}
       </div>