+++ themes/theme-b2b/src/components/Dialog.styles.ts
@@ -2,7 +2,7 @@ import { ThemeComponent, cva } from '@marigold/system';
 import { ELEVALTION_RING } from '../mixins';
 
 export const Dialog: ThemeComponent<'Dialog'> = {
-  closeButton: cva('absolute top-4 right-4 ml-4'),
+  closeButton: cva('absolute top-4 right-4 ml-4 size-4'),
   container: cva(
     [
       'font-body bg-bg-surface-overlay shadow-surface-overlay rounded-xs pl-8 pr-12 pb-8 pt-4',
@@ -23,6 +23,12 @@ export const Dialog: ThemeComponent<'Dialog'> = {
     }
   ),
   header: cva('flex items-center'),
+  title: cva('font-black text-3xl mb-1', {
+    variants: {
+      variant: {},
+      size: {},
+    },
+  }),
   content: cva(''),
-  actions: cva('flex gap-2 justify-end'),
+  actions: cva('flex gap-2 justify-end mt-4'),
 };