+++ themes/theme-rui/src/components/Dialog.styles.ts
@@ -1,12 +1,7 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const Dialog: ThemeComponent<'Dialog'> = {
-  closeButton: cva([
-    'absolute top-3 right-3',
-    'flex size-7 items-center justify-center rounded transition-[color,box-shadow]',
-    'mixin-ring-focus-visible',
-    '[&_svg]:size-4 [&_svg]:opacity-60 [&_svg]:transition-opacity hover:[&_svg]:opacity-100',
-  ]),
+  closeButton: cva(['absolute top-6 right-3', 'size-7 ']),
   container: cva(
     [
       'flex flex-col gap-0 p-0',
@@ -26,8 +21,14 @@ export const Dialog: ThemeComponent<'Dialog'> = {
     }
   ),
   header: cva('flex flex-col gap-1 text-center sm:text-left'),
+  title: cva('text-lg font-semibold mb-1', {
+    variants: {
+      variant: {},
+      size: {},
+    },
+  }),
   content: cva('text-muted-foreground text-sm'),
-  actions: cva('flex flex-col-reverse gap-3 sm:flex-row sm:justify-end'),
+  actions: cva('flex flex-col-reverse gap-3 sm:flex-row sm:justify-end mt-4'),
 };
 
 /**